1. RESTFul. Описание архитектуры [#6884]

+++ src/main/java/ru/job4j/auth/service/PersonService.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service;
 import ru.job4j.auth.domain.Person;
 import ru.job4j.auth.repository.PersonRepository;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,19 +14,29 @@ public class PersonService {
     private final PersonRepository personRepository;
 
     public Optional<Person> save(Person person) {
-        return Optional.of(personRepository.save(person));
-    }
-
-    public boolean deleteById(int id) {
-        boolean flag = false;
+        Optional<Person> saved = Optional.empty();
         try {
-            personRepository.deleteById(id);
-            flag = true;
+            saved = Optional.of(personRepository.save(person));
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return flag;
+        return saved;
+    }
 
+    public boolean deleteById(int id) {
+        boolean flag = findPersonById(id).isPresent();
+        if (flag) {
+            personRepository.deleteById(id);
+        }
+        return flag;
+    }
+
+    public boolean update(Person person) {
+        boolean flag = findPersonById(person.getId()).isPresent();
+        if (flag) {
+            personRepository.save(person);
+        }
+        return flag;
     }
 
     public Optional<Person> findPersonById(int id) {
